src/pagos/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
pydantic
pydantic-settings
pulsar-client[avro]
avro-python3
pytest
ruff
black
alembic
